+++ Utils/ProgressUtils.cs
@@ -7,11 +7,11 @@ namespace OrganizadorArquivosWPF.Utils
     /// </summary>
     public class ProgressTracker
     {
-        private readonly IProgress<int> _outer;
+        private readonly IProgress<double> _outer;
         private readonly int _totalSegments;
         private int _current;
 
-        public ProgressTracker(IProgress<int> outer, int totalSegments)
+        public ProgressTracker(IProgress<double> outer, int totalSegments)
         {
             _outer = outer;
             _totalSegments = Math.Max(1, totalSegments);
@@ -22,15 +22,15 @@ namespace OrganizadorArquivosWPF.Utils
         /// Gets a progress instance that maps its 0-100 range to the next segment
         /// of the outer progress.
         /// </summary>
-        public IProgress<int> NextSegment()
+        public IProgress<double> NextSegment()
         {
             int index = _current++;
-            int start = index * 100 / _totalSegments;
-            int end = (index + 1) * 100 / _totalSegments;
-            return new Progress<int>(v =>
+            double start = index * 100.0 / _totalSegments;
+            double end = (index + 1) * 100.0 / _totalSegments;
+            return new Progress<double>(v =>
             {
-                v = Math.Clamp(v, 0, 100);
-                int scaled = start + (v * (end - start) / 100);
+                v = Math.Clamp(v, 0.0, 100.0);
+                double scaled = start + (v * (end - start) / 100.0);
                 _outer.Report(scaled);
             });
         }
